feat: cbor and cleanup
