+++ packages/bls12381/package.json
@@ -34,6 +34,7 @@
     "@mattrglobal/bls12381-key-pair": "^0.4.1-unstable.256739c",
     "@transmute/did-key-common": "^0.2.1-unstable.28",
     "base64url": "^3.0.1",
+    "borc": "^2.1.2",
     "bs58": "^4.0.1"
   }
 }
