+++ packages/bls12381/src/driver.ts
@@ -2,7 +2,7 @@ import { getGet } from '@transmute/did-key-common';
 
 import { Bls12381KeyPairs } from './Bls12381KeyPairs';
 import { keyToDidDoc } from './functions/keyToDidDoc';
-
+const cbor = require('borc');
 export const getResolve = () => {
   const resolve = async (
     didUri: string,
@@ -13,14 +13,20 @@ export const getResolve = () => {
       .split('did:key:')
       .pop();
     const publicKey = await Bls12381KeyPairs.fromFingerprint({ fingerprint });
-    return {
+    const didResolutionResponse =  {
       didDocument: await keyToDidDoc(publicKey, resolutionMetaData.accept),
       didDocumentMetaData: {
         'content-type': resolutionMetaData.accept,
       },
       didResolutionMetaData: {},
     };
+    if (resolutionMetaData.accept === 'application/did+cbor'){
+      return cbor.encode(didResolutionResponse);
+    }
+    return didResolutionResponse;
   };
+
+  
   return resolve;
 };
