+++ packages/did-key-common/src/typings.d.ts
@@ -1,1 +1,2 @@
 declare module 'canonicalize';
+declare module 'cbor'
