+++ packages/ed25519/src/driver.test.ts
@@ -15,20 +15,25 @@ it('get interface defaults to application/did+ld+json', async () => {
 
 let representations = [
   {
-    keyType: 'JsonWebKey2020',
     contentType: 'application/did+json',
   },
   {
-    keyType: 'X25519KeyAgreementKey2019',
     contentType: 'application/did+ld+json',
+  },
+  {
+    contentType: 'application/did+cbor',
   },
 ];
 representations.forEach((rep) => {
-  let { id } = example.resolution[rep.contentType].didDocument;
+  let { id } = example.resolution['application/did+json'].didDocument;
   it(`resolve supports ${rep.contentType}`, async () => {
     let resolutionResponse: any = await resolve(id, {
       accept: rep.contentType,
     });
-    expect(resolutionResponse).toEqual(example.resolution[rep.contentType]);
+    if (rep.contentType === 'application/did+cbor'){
+      expect(Buffer.isBuffer(resolutionResponse)).toBe(true)
+    } else {
+      expect(resolutionResponse).toEqual(example.resolution[rep.contentType]);
+    }
   });
 });
