+++ packages/ed25519/src/typings.d.ts
@@ -1,3 +1,4 @@
 declare module 'bs58';
 declare module 'base64url';
 declare module 'canonicalize';
+declare module 'cbor-pretty-print'
