+++ packages/workbench/package.json
@@ -2,9 +2,9 @@
   "name": "@transmute/did-key-workbench",
   "version": "0.2.1-unstable.28",
   "private": true,
-  "homepage": "https://did-key.web.app",
+  "homepage": "https://did.key.transmute.industries",
   "dependencies": {
-    "@bloomprotocol/qr-react": "^2.0.0",
+
     "@material-ui/core": "^4.9.14",
     "@material-ui/icons": "^4.5.1",
     "@material-ui/lab": "^4.0.0-alpha.38",
@@ -17,10 +17,7 @@
     "@transmute/ed25519-signature-2018": "^0.1.1-unstable.6",
     "@transmute/json-web-signature-2020": "^1.1.0",
     "@transmute/jsonld-document-loader": "^0.2.0",
-    "@transmute/material-did-core": "0.0.8",
-    "@transmute/universal-wallet": "0.0.6",
     "@transmute/vc.js": "^0.1.1-unstable.6",
-    "bitcoin-ts": "1.14.2",
     "brace": "^0.11.1",
     "clsx": "^1.0.4",
     "connected-react-router": "^6.5.2",
@@ -36,24 +33,19 @@
     "libp2p-noise": "^2.0.1",
     "libp2p-webrtc-star": "^0.20.1",
     "libp2p-websockets": "^0.14.0",
-    "material-table": "^1.57.2",
+
     "moment": "^2.25.3",
-    "nom": "0.0.3",
-    "pako": "^1.0.11",
     "prop-types": "^15.7.2",
-    "qrcode.react": "^1.0.0",
+
     "react": "16.13.1",
     "react-ace": "^6.3.2",
     "react-dom": "16.13.1",
-    "react-force-graph-2d": "^1.16.0",
-    "react-identicons": "^1.2.4",
     "react-particles-js": "^3.2.1",
     "react-qr-reader": "^2.2.1",
     "react-redux": "^7.1.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.1",
     "react-swipeable-views": "^0.13.3",
-    "react-to-pdf": "0.0.11",
     "recompose": "^0.30.0",
     "redux": "^4.0.4",
     "redux-actions": "^2.6.5",
