--- packages/workbench/src/components/ScanQRCodeDialog/index.js
@@ -1,1 +1,0 @@
-export * from "./ScanQRCodeDialog";
