--- packages/workbench/src/pages/workbench/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Workbench as Page } from './Workbench';
-
-import { compose } from 'redux';
-
-import wallet from '../../store/universal-wallet';
-
-const container = compose(wallet.container);
-
-export const Workbench = container((props) => {
-  return <Page {...props} />;
-});
